requirements.txt
Django==5.2.1
django-q2
